Moved files from motion_interface modular to intera_interface and intrea_examples. The scripts now fully use the limb interface.
