+++ intera_examples/scripts/go_to_cartesian_pose.py
