+++ intera_examples/scripts/set_interaction_options.py
