+++ intera_interface/src/motion_interface/interaction_options.py
